Add files via upload
This project is done as a requirement of Computer Networks Course. This project deals with establishing communication between Vehicular Adhoc Networks using DABFS Algorithm.
VANET-Vehicular ad hoc networks is an intelligent component of the transport system that helps the vehicles to communicate with each other and with roadside components too. It plays a very important role in transportation systems by ensuring the fact that less accidents take place over a period of time. This vehicle-to-vehicle communication is a wireless transmission of data/messages between vehicles which are nearby and offers services for safety improvements. So, warning messages are transferred between the vehicles using a direction based greedy approach where the next hop is selected based on the current location of relay nodes, all the way towards the destination node.But this method works well with unidirectional traffic.Since the vehicles are not permitted to have unidirectional traffic,this method proved to be ineffective because of more hops and end-to-end delay.

So,This project deals with the positions of source and destination nodes in a Bidirectional environment and discusses its complexity further for efficient and robust routing of warning messages using a routing protocol namely, Direction aware Best Forward Selection(DABFS).
DABFS takes into account directions and relative positions of nodes along with the distance parameter to determine a node’s movement direction using Hamming distance and forwards the messages through neighbors and discovers the best route to reach the destination. This proves to be the efficient protocol that increases the throughput and reduces packet loss along with end-to-end delay